refactor(internal): replace runners.js by ioRunner.js
and move createTestRunner code directly in testHandler implementation

Closes #31
