+++ src/internal/ioRunner.js
@@ -1,0 +1,3 @@
+const ioRunner = ({ f, args }) => f(...args);
+
+export default ioRunner;
